requirements.txt
torch>=2.0.0
diffusers>=0.20.0
transformers>=4.25.0
torchvision>=0.15.0
tqdm
clip-by-openai
scipy